Trait relaxed_pcs::IF

pub trait IF: Field {
    type Values: IntoIterator<Item = Self>;

    fn all_values() -> Self::Values;

    fn all_multidimentional_values(m: usize) -> Vec<Vec<Self>> { ... }
}

Iterate over all possible values of a finite field.

Required Associated Types

type Values: IntoIterator<Item = Self>

Type of the values.

Required Methods

fn all_values() -> Self::Values

Get all values of the type.

Provided Methods

fn all_multidimentional_values(m: usize) -> Vec<Vec<Self>>

Get all permutations of the values of the type.